frontend/tests/e2e/requirements.txt
pytest>=8.0.0
playwright>=1.48.0
pyxsql>=1.7.0
